requirements.txt
requests~=2.27
python-telegram-bot~=13.11
python-dotenv~=0.20
Flask~=2.1
gunicorn~=20.1